refactor: Flyttet Input controls ut i egen komponent
